requirements.txt
aiosmtplib==5.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
limits==5.8.0
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.42.0
vine==5.1.0
wcwidth==0.6.0
wrapt==2.1.2
